DeepRisk-analyze-server/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.30.6
nacos-sdk-python==0.1.8
redis==5.0.8
pika==1.3.2
numpy==1.24.3
scikit-learn==1.3.0
torch==2.1.0
torchvision==0.16.0
pandas==2.0.3
pyyaml==6.0.1
requests==2.31.0
websocket-client==1.8.0
aiohttp==3.9.5
mysql-connector-python==8.0.33
pymysql==1.1.0
ray==2.8.0
einops==0.7.0
tqdm==4.66.1
scipy==1.11.3
pydantic==2.5.0